requirements.txt
pytelegrambotapi==4.16.1
python-dotenv==1.0.0
requests==2.31.0
wakeonlan==3.1.0
